+++ sc-memory/sc-memory/sc_keynodes.hpp
@@ -22,7 +22,9 @@ private:
 
 public:
   // Just for internal usage
-  static _SC_EXTERN bool Init(bool force = false, sc_char const * result_structure_system_idtf = (sc_char *)null_ptr);
+  static _SC_EXTERN bool Init(
+      bool force = false,
+      sc_char const * init_memory_generated_structure = (sc_char *)null_ptr);
   static void Shutdown();
 
   static _SC_EXTERN ScAddr const & GetResultCodeAddr(sc_result resCode);