+++ sc-tools/sc-builder/src/main.cpp
@@ -48,8 +48,13 @@ try
   ScParams memoryParams{options, keys};
   memoryParams.insert({"repo_path", options[{"output_path", "o"}].second});
 
-  ScMemoryConfig memoryConfig{config, {"repo_path", "log_file"}, std::move(memoryParams)};
   ScBuilderConfig builderConfig{ScConfig(config, {"repo_path", "log_file"}), std::move(params)};
+
+  memoryParams.insert(
+      {"init_memory_generated_upload", std::to_string(builderConfig.GetParams().m_resultStructureUpload)});
+  memoryParams.insert({"init_memory_generated_structure", builderConfig.GetParams().m_resultStructureSystemIdtf});
+
+  ScMemoryConfig memoryConfig{config, {"repo_path", "log_file"}, std::move(memoryParams)};
 
   Builder builder;
   return builder.Run(builderConfig.GetParams(), memoryConfig.GetParams()) ? EXIT_SUCCESS : EXIT_FAILURE;