+++ docs/changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Insert sc-keynodes in global knowledge base during sc-memory initiation 
 - CI for documentation build
 - Fix memory clear parameter reading from config file
 - Add output structure builder feature to config